Engine: Introduce CardSpecification abstraction
For alternative specification implementations
